Change buildrun create command to have random name
There could be the use case that a user will run multiple buildruns
while due to frequent changes during development cycles. It can be
cumbersome to come up with a unused name for the buildrun every time.

Change code to accept either one or no buildrun name. In case none is
provided, come up with a random name for the buildrun.