SHIP-0038: Add Git Ref to Release Workflow
Update the release workflow to accept a git reference as a parameter.
This will be used to check out the appropriate release branch when
generating artifacts and release notes. The workflow then creates a tag
for the release, which is required by GoReleaser in a subsequent step.

The action was also updated to use the (new?) inputs context, instead
of relying on the inputs key existing in the github.events payload.